+++ extensions/frmtmb.eam/R/wiener-gng.R
@@ -559,7 +559,7 @@ gng_family <- function(cfg, ub, delta) {
 
   fam <- frmtmb::custom_family(
     "wiener_gng",
-    accepts_aterms = c("dec", "vreal", "weights", "cens"),
+    accepts_aterms = ddm_accepts[["wiener_gng"]],
     dpars = dpars,
     links = links,
     lpdf = lpdf,